Exposed services code refactor
