+++ pkg/backends/k8s.go
@@ -24,7 +24,6 @@ import (
 	"github.com/goccy/go-yaml"
 	"github.com/grycap/oscar/v3/pkg/imagepuller"
 	"github.com/grycap/oscar/v3/pkg/types"
-	"github.com/grycap/oscar/v3/pkg/utils"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -77,8 +76,6 @@ func (k *KubeBackend) ListServices() ([]*types.Service, error) {
 
 // CreateService creates a new service as a k8s podTemplate
 func (k *KubeBackend) CreateService(service types.Service) error {
-	// Validate the input variables of the service
-	service = utils.ValidateService(service)
 	// Create the configMap with FDL and user-script
 	err := createServiceConfigMap(&service, k.namespace, k.kubeClientset)
 	if err != nil {
@@ -117,8 +114,8 @@ func (k *KubeBackend) CreateService(service types.Service) error {
 	}
 
 	//Create an expose service
-	if service.Expose.Port != 0 {
-		utils.CreateExpose(service, k.kubeClientset, k.config)
+	if service.Expose.APIPort != 0 {
+		types.CreateExpose(service, k.kubeClientset, k.config)
 	}
 	//Create deaemonset to cache the service image on all the nodes
 	if service.ImagePrefetch {
@@ -149,8 +146,6 @@ func (k *KubeBackend) ReadService(name string) (*types.Service, error) {
 
 // UpdateService updates an existent service
 func (k *KubeBackend) UpdateService(service types.Service) error {
-	// Validate the input variables of the service
-	service = utils.ValidateService(service)
 	// Get the old service's configMap
 	oldCm, err := k.kubeClientset.CoreV1().ConfigMaps(k.namespace).Get(context.TODO(), service.Name, metav1.GetOptions{})
 	if err != nil {
@@ -196,14 +191,20 @@ func (k *KubeBackend) UpdateService(service types.Service) error {
 		return err
 	}
 
-	//Update an expose service
-	utils.UpdateExpose(service, k.kubeClientset, k.config)
+	// If the service is exposed update its configuration
+	if service.Expose.APIPort != 0 {
+		err = types.UpdateExpose(service, k.kubeClientset, k.config)
+		if err != nil {
+			return err
+		}
+	}
 
 	return nil
 }
 
 // DeleteService deletes a service
-func (k *KubeBackend) DeleteService(name string) error {
+func (k *KubeBackend) DeleteService(service types.Service) error {
+	name := service.Name
 	if err := k.kubeClientset.CoreV1().PodTemplates(k.namespace).Delete(context.TODO(), name, metav1.DeleteOptions{}); err != nil {
 		return err
 	}
@@ -218,9 +219,13 @@ func (k *KubeBackend) DeleteService(name string) error {
 		log.Printf("Error deleting associated jobs for service \"%s\": %v\n", name, err)
 	}
 
-	if err2 := utils.DeleteExpose(name, k.kubeClientset, k.config); err2 != nil {
-		log.Printf("Error deleting all associated kubernetes component of an exposed service \"%s\": %v\n", name, err2)
+	// If service is exposed delete the exposed k8s components
+	if service.Expose.APIPort != 0 {
+		if err := types.DeleteExpose(name, k.kubeClientset, k.config); err != nil {
+			log.Printf("Error deleting all associated kubernetes component of the expose config \"%s\": %v\n", name, err)
+		}
 	}
+
 	return nil
 }
