+++ pkg/backends/knative.go
@@ -26,7 +26,6 @@ import (
 
 	"github.com/grycap/oscar/v3/pkg/imagepuller"
 	"github.com/grycap/oscar/v3/pkg/types"
-	"github.com/grycap/oscar/v3/pkg/utils"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -98,8 +97,6 @@ func (kn *KnativeBackend) ListServices() ([]*types.Service, error) {
 
 // CreateService creates a new service as a Knative service
 func (kn *KnativeBackend) CreateService(service types.Service) error {
-	// Validate the input variables of the service
-	service = utils.ValidateService(service)
 	// Create the configMap with FDL and user-script
 	err := createServiceConfigMap(&service, kn.namespace, kn.kubeClientset)
 	if err != nil {
@@ -127,8 +124,8 @@ func (kn *KnativeBackend) CreateService(service types.Service) error {
 	}
 
 	//Create an expose service
-	if service.Expose.Port != 0 {
-		utils.CreateExpose(service, kn.kubeClientset, kn.config)
+	if service.Expose.APIPort != 0 {
+		types.CreateExpose(service, kn.kubeClientset, kn.config)
 	}
 	//Create deaemonset to cache the service image on all the nodes
 	if service.ImagePrefetch {
@@ -164,8 +161,6 @@ func (kn *KnativeBackend) UpdateService(service types.Service) error {
 	if err != nil {
 		return err
 	}
-	// Validate the input variables of the service
-	service = utils.ValidateService(service)
 	// Get the old service's configMap
 	oldCm, err := kn.kubeClientset.CoreV1().ConfigMaps(kn.namespace).Get(context.TODO(), service.Name, metav1.GetOptions{})
 	if err != nil {
@@ -206,18 +201,25 @@ func (kn *KnativeBackend) UpdateService(service types.Service) error {
 		}
 		return err
 	}
-	//Update an expose service
-	utils.UpdateExpose(service, kn.kubeClientset, kn.config)
+
+	// If the service is exposed update its configuration
+	if service.Expose.APIPort != 0 {
+		err = types.UpdateExpose(service, kn.kubeClientset, kn.config)
+		if err != nil {
+			return err
+		}
+	}
 
 	return nil
 }
 
 // DeleteService deletes a service
-func (kn *KnativeBackend) DeleteService(name string) error {
+func (kn *KnativeBackend) DeleteService(service types.Service) error {
+
+	name := service.Name
 	if err := kn.knClientset.ServingV1().Services(kn.namespace).Delete(context.TODO(), name, metav1.DeleteOptions{}); err != nil {
 		return err
 	}
-
 	// Delete the service's configMap
 	if delErr := deleteServiceConfigMap(name, kn.namespace, kn.kubeClientset); delErr != nil {
 		log.Println(delErr.Error())
@@ -227,8 +229,12 @@ func (kn *KnativeBackend) DeleteService(name string) error {
 	if err := deleteServiceJobs(name, kn.namespace, kn.kubeClientset); err != nil {
 		log.Printf("Error deleting associated jobs for service \"%s\": %v\n", name, err)
 	}
-	if err2 := utils.DeleteExpose(name, kn.kubeClientset, kn.config); err2 != nil {
-		log.Printf("Error deleting all associated kubernetes component of an exposed service \"%s\": %v\n", name, err2)
+
+	// If service is exposed delete the exposed k8s components
+	if service.Expose.APIPort != 0 {
+		if err := types.DeleteExpose(name, kn.kubeClientset, kn.config); err != nil {
+			log.Printf("Error deleting all associated kubernetes component of an exposed service \"%s\": %v\n", name, err)
+		}
 	}
 
 	return nil
