+++ pkg/types/config.go
@@ -181,7 +181,8 @@ type Config struct {
 
 	//
 	IngressHost string `json:"-"`
-	//
+
+	// Github path of FaaS Supervisor (needed for Interlink config)
 	SupervisorURL string `json:"-"`
 }
 
@@ -361,6 +362,7 @@ func (cfg *Config) CheckAvailableGPUs(kubeClientset kubernetes.Interface) {
 	}
 }
 
+// CheckAvailableInterLink checks if there is a node with the virtual kubelet annotation
 func (cfg *Config) CheckAvailableInterLink(kubeClientset kubernetes.Interface) {
 	nodes, err := kubeClientset.CoreV1().Nodes().List(context.TODO(), metav1.ListOptions{LabelSelector: "!node-role.kubernetes.io/control-plane,!node-role.kubernetes.io/master,type=virtual-kubelet"})
 	if err != nil {
