+++ pkg/types/expose.go
@@ -1,0 +1,645 @@
+/*
+Copyright (C) GRyCAP - I3M - UPV
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package types
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+
+	htpasswd "github.com/foomo/htpasswd"
+	apps "k8s.io/api/apps/v1"
+	autos "k8s.io/api/autoscaling/v1"
+	v1 "k8s.io/api/core/v1"
+	net "k8s.io/api/networking/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	"k8s.io/client-go/kubernetes"
+)
+
+const (
+	typeNodePort      = "NodePort"
+	typeClusterIP     = "ClusterIP"
+	prefixLabelApp    = "oscar-svc-exp-"
+	keyLabelApp       = "app"
+	podPortName       = "podport"
+	servicePortName   = "serviceport"
+	servicePortNumber = 80
+)
+
+// Custom logger
+var ExposeLogger = log.New(os.Stdout, "[EXPOSED-SERVICE] ", log.Flags())
+
+/* Exposed service is composed by a deployment and a service.
+An exposed service can be of to types:
+- NodePort
+- Ingress */
+
+// CreateExpose creates all the kubernetes components
+func CreateExpose(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	ExposeLogger.Printf("Creating exposed service: \n%v\n", service)
+	err := createDeployment(service, kubeClientset, cfg)
+	if err != nil {
+		return fmt.Errorf("error creating deployment for exposed service '%s': %v", service.Name, err)
+	}
+	err = createService(service, kubeClientset, cfg)
+	if err != nil {
+		return fmt.Errorf("error creating svc for exposed service '%s': %v", service.Name, err)
+	}
+	if service.Expose.NodePort == 0 {
+		err = createIngress(service, kubeClientset, cfg)
+		if err != nil {
+			return fmt.Errorf("error creating ingress for exposed service '%s': %v", service.Name, err)
+		}
+	}
+	return nil
+}
+
+// DeleteExpose removes all the components of the exposed service from the cluster
+func DeleteExpose(name string, kubeClientset kubernetes.Interface, cfg *Config) error {
+	err := deleteDeployment(name, kubeClientset, cfg)
+	if err != nil {
+		return fmt.Errorf("error deleting deployment for exposed service '%s': %v", name, err)
+	}
+	err = deleteService(name, kubeClientset, cfg)
+	if err != nil {
+		return fmt.Errorf("error deleting service for exposed service '%s': %v", name, err)
+	}
+
+	ingressType := existsIngress(name, cfg.Namespace, kubeClientset)
+	if ingressType {
+		err = deleteIngress(name, kubeClientset, cfg)
+		if err != nil {
+			return fmt.Errorf("error deleting ingress for exposed service '%s': %v", name, err)
+		}
+	}
+
+	return nil
+}
+
+// UpdateExpose updates all the components of the exposed service on the cluster
+func UpdateExpose(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+
+	// If the deployment exist, and keep continue it will upload.
+	err := updateDeployment(service, kubeClientset, cfg)
+	if err != nil {
+		ExposeLogger.Printf("error updating exposed service deployment: %v\n", err)
+		return err
+	}
+	err = updateService(service, kubeClientset, cfg)
+	if err != nil {
+		ExposeLogger.Printf("error updating exposed service: %v\n", err)
+		return err
+	}
+
+	ingressType := existsIngress(service.Name, cfg.Namespace, kubeClientset)
+	// Old service config was Ingress type
+	if ingressType {
+		// New service config if NodePort
+		if service.Expose.NodePort != 0 {
+			err = deleteIngress(getIngressName(service.Name), kubeClientset, cfg)
+			if err != nil {
+				log.Printf("error deleting ingress service: %v\n", err)
+				return err
+			} else {
+				err = updateIngress(service, kubeClientset, cfg)
+				if err != nil {
+					log.Printf("error updating ingress service: %v\n", err)
+					return err
+				}
+			}
+		}
+	} else {
+		// Old service config is NodeType and the new one is Ingress type
+		if service.Expose.NodePort == 0 {
+			err = createIngress(service, kubeClientset, cfg)
+			if err != nil {
+				log.Printf("error creating ingress service: %v\n", err)
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+// TODO check and refactor
+// Main function that list all the kubernetes components
+// This function is not used, in the future could be usefull
+func ListExpose(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	deploy, hpa, err := listDeployments(kubeClientset, cfg)
+
+	services, err2 := listServices(kubeClientset, cfg)
+	ingress, err3 := listIngress(kubeClientset, cfg)
+	if err != nil {
+		ExposeLogger.Printf("WARNING: %v\n", err)
+		return err
+	}
+	if err2 != nil {
+		ExposeLogger.Printf("WARNING: %v\n", err2)
+		return err
+	}
+	if err3 != nil {
+		ExposeLogger.Printf("WARNING: %v\n", err3)
+		return err
+	}
+	fmt.Println(deploy, hpa, services, ingress)
+	return nil
+
+}
+
+//////////// Deployment
+
+/// Create deployment and horizontal autoscale
+
+func createDeployment(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	deployment := getDeploymentSpec(service, cfg)
+	_, err := kubeClientset.AppsV1().Deployments(cfg.ServicesNamespace).Create(context.TODO(), deployment, metav1.CreateOptions{})
+	if err != nil {
+		return err
+	}
+
+	hpa := getHortizontalAutoScaleSpec(service, cfg)
+	_, err = kubeClientset.AutoscalingV1().HorizontalPodAutoscalers(cfg.ServicesNamespace).Create(context.TODO(), hpa, metav1.CreateOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// Return the component deployment, ready to create or update
+func getDeploymentSpec(service Service, cfg *Config) *apps.Deployment {
+	deployName := getDeploymentName(service.Name)
+	deployment := &apps.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      deployName,
+			Namespace: cfg.ServicesNamespace,
+		},
+		Spec: apps.DeploymentSpec{
+			Replicas: &service.Expose.MinScale,
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					keyLabelApp: prefixLabelApp + service.Name,
+				},
+			},
+			Template: getPodTemplateSpec(service, cfg),
+		},
+		Status: apps.DeploymentStatus{},
+	}
+
+	return deployment
+}
+
+// Return the component HorizontalAutoScale, ready to create or update
+func getHortizontalAutoScaleSpec(service Service, cfg *Config) *autos.HorizontalPodAutoscaler {
+	hpaName := getHPAName(service.Name)
+	deployName := getDeploymentName(service.Name)
+	hpa := &autos.HorizontalPodAutoscaler{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      hpaName,
+			Namespace: cfg.ServicesNamespace,
+		},
+		Spec: autos.HorizontalPodAutoscalerSpec{
+			ScaleTargetRef: autos.CrossVersionObjectReference{
+				Kind:       "Deployment",
+				Name:       deployName,
+				APIVersion: "apps/v1",
+			},
+			MinReplicas:                    &service.Expose.MinScale,
+			MaxReplicas:                    service.Expose.MaxScale,
+			TargetCPUUtilizationPercentage: &service.Expose.CpuThreshold,
+		},
+		Status: autos.HorizontalPodAutoscalerStatus{},
+	}
+	return hpa
+}
+
+// Return the Pod spec inside of deployment, ready to create or update
+
+func getPodTemplateSpec(service Service, cfg *Config) v1.PodTemplateSpec {
+	podSpec, _ := service.ToPodSpec(cfg)
+
+	for i, _ := range podSpec.Containers {
+		podSpec.Containers[i].Ports = []v1.ContainerPort{
+			{
+				Name:          podPortName,
+				ContainerPort: int32(service.Expose.APIPort),
+			},
+		}
+		podSpec.Containers[i].Resources = v1.ResourceRequirements{
+			Requests: v1.ResourceList{
+				"cpu": *resource.NewMilliQuantity(500, resource.DecimalSI),
+			},
+		}
+		podSpec.Containers[i].VolumeMounts[0].ReadOnly = false
+		if service.Expose.DefaultCommand {
+			podSpec.Containers[i].Command = nil
+			podSpec.Containers[i].Args = nil
+		} else {
+			podSpec.Containers[i].Command = []string{"/bin/sh"}
+			podSpec.Containers[i].Args = []string{"-c", fmt.Sprintf("%s/%s", ConfigPath, ScriptFileName)}
+		}
+	}
+	var num int32 = 0777
+	podSpec.Volumes[0].VolumeSource.ConfigMap.DefaultMode = &num
+	template := v1.PodTemplateSpec{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      service.Name,
+			Namespace: service.Image,
+			Labels: map[string]string{
+				keyLabelApp: prefixLabelApp + service.Name,
+			},
+		},
+		Spec: *podSpec,
+	}
+	return template
+}
+
+// / List deployment and the horizontal auto scale
+func listDeployments(kubeClientset kubernetes.Interface, cfg *Config) (*apps.DeploymentList, *autos.HorizontalPodAutoscalerList, error) {
+	deployment, err := kubeClientset.AppsV1().Deployments(cfg.ServicesNamespace).List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		return nil, nil, err
+	}
+
+	hpa, err2 := kubeClientset.AutoscalingV1().HorizontalPodAutoscalers(cfg.ServicesNamespace).List(context.TODO(), metav1.ListOptions{})
+	if err2 != nil {
+		return nil, nil, err2
+	}
+	return deployment, hpa, nil
+}
+
+// Delete Deployment and HPA
+func deleteDeployment(name string, kubeClientset kubernetes.Interface, cfg *Config) error {
+	name_hpa := getHPAName(name)
+	err := kubeClientset.AutoscalingV1().HorizontalPodAutoscalers(cfg.ServicesNamespace).Delete(context.TODO(), name_hpa, metav1.DeleteOptions{})
+	if err != nil {
+		return err
+	}
+	deployment := getDeploymentName(name)
+	err = kubeClientset.AppsV1().Deployments(cfg.ServicesNamespace).Delete(context.TODO(), deployment, metav1.DeleteOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+///Update Deployment and HPA
+
+func updateDeployment(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	_, err := kubeClientset.AppsV1().Deployments(cfg.ServicesNamespace).Get(context.TODO(), getDeploymentName(service.Name), metav1.GetOptions{})
+	if err != nil {
+		return err
+	}
+
+	deployment := getDeploymentSpec(service, cfg)
+	_, err = kubeClientset.AppsV1().Deployments(cfg.ServicesNamespace).Update(context.TODO(), deployment, metav1.UpdateOptions{})
+	if err != nil {
+		return err
+	}
+
+	kubeClientset.AutoscalingV1().HorizontalPodAutoscalers(cfg.ServicesNamespace).Get(context.TODO(), getHPAName(service.Name), metav1.GetOptions{})
+	hpa := getHortizontalAutoScaleSpec(service, cfg)
+	_, err = kubeClientset.AutoscalingV1().HorizontalPodAutoscalers(cfg.ServicesNamespace).Update(context.TODO(), hpa, metav1.UpdateOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+/////////// Service
+
+// Create a kubernetes service component
+func createService(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	service_spec := getServiceSpec(service, cfg)
+	_, err := kubeClientset.CoreV1().Services(cfg.ServicesNamespace).Create(context.TODO(), service_spec, metav1.CreateOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// Return a kubernetes service component, ready to deploy or update
+func getServiceSpec(service Service, cfg *Config) *v1.Service {
+	name_service := getServiceName(service.Name)
+	var port v1.ServicePort = v1.ServicePort{
+		Name: servicePortName,
+		Port: servicePortNumber,
+		TargetPort: intstr.IntOrString{
+			Type:   0,
+			IntVal: int32(service.Expose.APIPort),
+		},
+	}
+	service_type := v1.ServiceType(typeClusterIP)
+	if service.Expose.NodePort != 0 {
+		service_type = typeNodePort
+		port.NodePort = service.Expose.NodePort
+	}
+	service_spec := &v1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name_service,
+			Namespace: cfg.ServicesNamespace,
+		},
+		Spec: v1.ServiceSpec{
+			Ports: []v1.ServicePort{port},
+			Type:  service_type,
+			Selector: map[string]string{
+				keyLabelApp: prefixLabelApp + service.Name,
+			},
+		},
+		Status: v1.ServiceStatus{},
+	}
+	return service_spec
+}
+
+/// List services in a certain namespace
+
+func listServices(kubeClientset kubernetes.Interface, cfg *Config) (*v1.ServiceList, error) {
+	services, err := kubeClientset.CoreV1().Services(cfg.ServicesNamespace).List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
+	return services, nil
+}
+
+// / Update a kubernete service
+func updateService(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	kube_service := getServiceSpec(service, cfg)
+	_, err := kubeClientset.CoreV1().Services(cfg.ServicesNamespace).Update(context.TODO(), kube_service, metav1.UpdateOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+/// Delete kubernetes service
+
+func deleteService(name string, kubeClientset kubernetes.Interface, cfg *Config) error {
+	kube_service := getServiceName(name)
+	err := kubeClientset.CoreV1().Services(cfg.ServicesNamespace).Delete(context.TODO(), kube_service, metav1.DeleteOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+/////////// Ingress
+
+// / Create an ingress component
+func createIngress(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	// Create Secret
+
+	ingress := getIngressSpec(service, kubeClientset, cfg)
+	_, err := kubeClientset.NetworkingV1().Ingresses(cfg.ServicesNamespace).Create(context.TODO(), ingress, metav1.CreateOptions{})
+	if err != nil {
+		return err
+	}
+	if service.Expose.SetAuth {
+		createSecret(service, kubeClientset, cfg)
+	}
+	return nil
+}
+
+// / Update a kubernete service
+func updateIngress(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+
+	serviceName := service.Name
+	//if exist continue and need -> Update
+	//if exist and not need -> delete
+	//if not  exist create
+	kube_ingress := getIngressSpec(service, kubeClientset, cfg)
+	_, err := kubeClientset.NetworkingV1().Ingresses(cfg.ServicesNamespace).Update(context.TODO(), kube_ingress, metav1.UpdateOptions{})
+	if err != nil {
+		return err
+	}
+
+	secret := existsSecret(serviceName, kubeClientset, cfg)
+	if secret {
+		if service.Expose.SetAuth {
+			updateSecret(service, kubeClientset, cfg)
+		} else {
+			deleteSecret(service.Name, kubeClientset, cfg)
+		}
+	} else {
+		if service.Expose.SetAuth {
+			createSecret(service, kubeClientset, cfg)
+		}
+	}
+
+	return nil
+}
+
+// Return a kubernetes ingress component, ready to deploy or update
+func getIngressSpec(service Service, kubeClientset kubernetes.Interface, cfg *Config) *net.Ingress {
+	name_ingress := getIngressName(service.Name)
+	pathofapi := getAPIPath(service.Name)
+	name_service := getServiceName(service.Name)
+	var ptype net.PathType = "Prefix"
+	var ingresspath net.HTTPIngressPath = net.HTTPIngressPath{
+		Path:     pathofapi + "/?(.*)",
+		PathType: &ptype,
+		Backend: net.IngressBackend{
+			Service: &net.IngressServiceBackend{
+				Name: name_service,
+				Port: net.ServiceBackendPort{
+					Number: servicePortNumber,
+				},
+			},
+		},
+	}
+	var ingresssrulevalue net.HTTPIngressRuleValue = net.HTTPIngressRuleValue{
+		Paths: []net.HTTPIngressPath{ingresspath},
+	}
+	var rule net.IngressRule
+	var tls net.IngressTLS
+	var specification net.IngressSpec
+
+	var host string = cfg.IngressHost
+
+	if host == "" {
+		rule = net.IngressRule{
+			IngressRuleValue: net.IngressRuleValue{HTTP: &ingresssrulevalue},
+		}
+		specification = net.IngressSpec{
+			TLS:   []net.IngressTLS{},
+			Rules: []net.IngressRule{rule}, //IngressClassName:
+		}
+	} else {
+		rule = net.IngressRule{
+			Host:             host,
+			IngressRuleValue: net.IngressRuleValue{HTTP: &ingresssrulevalue},
+		}
+		tls = net.IngressTLS{
+			Hosts:      []string{host},
+			SecretName: host,
+		}
+		specification = net.IngressSpec{
+			TLS:   []net.IngressTLS{tls},
+			Rules: []net.IngressRule{rule}, //IngressClassName:
+		}
+	}
+
+	rewriteOption := "/$1"
+	if service.Expose.RewriteTarget {
+		rewriteOption = pathofapi + "/$1"
+	}
+	annotation := map[string]string{
+		"nginx.ingress.kubernetes.io/rewrite-target": rewriteOption,
+		"kubernetes.io/ingress.class":                "nginx",
+		"nginx.ingress.kubernetes.io/use-regex":      "true",
+	}
+	if service.Expose.SetAuth {
+		annotation["nginx.ingress.kubernetes.io/auth-type"] = "basic"
+		annotation["nginx.ingress.kubernetes.io/auth-secret"] = getSecretName(service.Name)
+		annotation["nginx.ingress.kubernetes.io/auth-realm"] = "Authentication Required"
+	}
+	ingress := &net.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        name_ingress,
+			Namespace:   cfg.ServicesNamespace,
+			Annotations: annotation,
+		},
+		Spec:   specification,
+		Status: net.IngressStatus{},
+	}
+	return ingress
+}
+
+/// List the kuberntes ingress
+
+func listIngress(kubeClientset kubernetes.Interface, cfg *Config) (*net.IngressList, error) {
+	ingress, err := kubeClientset.NetworkingV1().Ingresses(cfg.ServicesNamespace).List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
+	return ingress, nil
+}
+
+// Delete a kubernetes ingress
+func deleteIngress(name string, kubeClientset kubernetes.Interface, cfg *Config) error {
+	// if secret exist, delete
+	ingress := getIngressName(name)
+	err := kubeClientset.NetworkingV1().Ingresses(cfg.ServicesNamespace).Delete(context.TODO(), ingress, metav1.DeleteOptions{})
+	if err != nil {
+		return err
+	}
+	deleteSecret(name, kubeClientset, cfg)
+	return nil
+}
+
+// Secret
+
+func createSecret(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	secret := getSecretSpec(service, kubeClientset, cfg)
+	_, err := kubeClientset.CoreV1().Secrets(cfg.ServicesNamespace).Create(context.TODO(), secret, metav1.CreateOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func updateSecret(service Service, kubeClientset kubernetes.Interface, cfg *Config) error {
+	secret := getSecretSpec(service, kubeClientset, cfg)
+	_, err := kubeClientset.CoreV1().Secrets(cfg.ServicesNamespace).Update(context.TODO(), secret, metav1.UpdateOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func deleteSecret(name string, kubeClientset kubernetes.Interface, cfg *Config) error {
+	secret := getSecretName(name)
+	err := kubeClientset.CoreV1().Secrets(cfg.ServicesNamespace).Delete(context.TODO(), secret, metav1.DeleteOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+func getSecretSpec(service Service, kubeClientset kubernetes.Interface, cfg *Config) *v1.Secret {
+	//setPassword
+	hash := make(htpasswd.HashedPasswords)
+	err := hash.SetPassword(service.Name, service.Token, htpasswd.HashAPR1)
+	if err != nil {
+		ExposeLogger.Printf(err.Error())
+	}
+	//Create Secret
+	inmutable := false
+	secret := &v1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      getSecretName(service.Name),
+			Namespace: cfg.ServicesNamespace,
+		},
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Secret",
+			APIVersion: "v1",
+		},
+		Immutable: &inmutable,
+		StringData: map[string]string{
+			"auth": service.Name + ":" + hash[service.Name],
+		},
+		Type: "Opaque",
+	}
+	return secret
+}
+
+func existsSecret(serviceName string, kubeClientset kubernetes.Interface, cfg *Config) bool {
+	secret := getSecretName(serviceName)
+	exist, err := kubeClientset.CoreV1().Secrets(cfg.ServicesNamespace).Get(context.TODO(), secret, metav1.GetOptions{})
+	if err != nil {
+		return false
+	}
+	if exist != nil {
+		return true
+	}
+	return false
+}
+
+func existsIngress(serviceName string, namespace string, kubeClientset kubernetes.Interface) bool {
+	_, err := kubeClientset.NetworkingV1().Ingresses(namespace).Get(context.TODO(), serviceName, metav1.GetOptions{})
+	if err == nil {
+		return true
+	}
+	return false
+}
+
+/// These are auxiliary functions
+
+func getServiceName(name_container string) string {
+	return name_container + "-svc"
+}
+
+func getIngressName(name_container string) string {
+	return name_container + "-ing"
+}
+
+func getAPIPath(name_container string) string {
+	return "/system/services/" + name_container + "/exposed"
+}
+
+func getDeploymentName(name_container string) string {
+	return name_container + "-dlp"
+}
+
+func getHPAName(name_container string) string {
+	return name_container + "-hpa"
+}
+
+func getSecretName(name_container string) string {
+	return name_container + "-auth-expose"
+}
